move reference utils from utilities package to core package
