+++ packages/core/src/index.ts
@@ -11,3 +11,4 @@ export * from "./types.js"
 export * from "./utils/index.js"
 export * from "./version.js"
 export * from "./m3/index.js"
+export * from "./m1/reference-utils.js"
