+++ packages/utilities/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Upgrade to `littoral-templates` version 0.3.0.
 * (Depend on `@lionweb/core` version 0.6.10.)
+* Move the reference util functions (`referenceValues`, `incomingReferences`, `referencesToOutOfScopeNodes`) and type (class `ReferenceValue`) from the `utilities` package to `core`.
+    Aliases (as re-exports) are left in its place to avoid code breaking.
+    The move is necessary because not all clients of `@lionweb/utilities` can install the Node.js-specific dependencies of the dependencies that package (such as `crypto`).
 
 
 ## 0.6.9
